connection changes to not happen automated when GUI starts, but only when login button is pressed. -> clien instaciated. notice that you MUST enter ip/port/username before pressin login!
feature added: user list, that updates everytime we recieve CLIENTLIST:

users in userList can be selected, messages will now ONLY be sent to selected users, if no users selected we send to all

send methods are seperated and changed.

dispatch on MSGRES or CLIENTLIST is changed to better logic and easier readability. recieved clientlist is updated to be splitted and strings are now generated using .append from a stingwriter.